Третье задание перенесено в отдельный класс
